refactor(session): colocate run actions with session flow
